Former-commit-id: f7c8c8d30d1cbb4b656aafe6723759c4555453e0

+++ trunk/src/protocol/metadata/downloadfilerequest.cc
@@ -137,7 +137,7 @@ void DownloadFileRequestMsg::setSize (uint64_t size) {
 	_size = size;
 }
 
-uint32_t DownloadFileRequestMsg::getSize() {
+uint64_t DownloadFileRequestMsg::getSize() {
 	return _size;
 }
